Move methods used only once to private scope
