+++ src/dotnet/commands/dotnet-sln/add/Program.cs
@@ -64,7 +64,7 @@ namespace Microsoft.DotNet.Tools.Sln.Add
                         PathUtility.EnsureTrailingSlash(slnFile.BaseDirectory),
                         fullProjectPath);
 
-                    return SlnProjectExtensions.GetSolutionFoldersFromProjectPath(relativeProjectPath);
+                    return GetSolutionFoldersFromProjectPath(relativeProjectPath);
                 };
             }
         }
@@ -104,6 +104,47 @@ namespace Microsoft.DotNet.Tools.Sln.Add
             }
 
             return 0;
+        }
+
+        private static IList<string> GetSolutionFoldersFromProjectPath(string projectFilePath)
+        {
+            var solutionFolders = new List<string>();
+
+            if (IsPathInTreeRootedAtSolutionDirectory(projectFilePath))
+            {
+                var currentDirString = $".{Path.DirectorySeparatorChar}";
+                if (projectFilePath.StartsWith(currentDirString))
+                {
+                    projectFilePath = projectFilePath.Substring(currentDirString.Length);
+                }
+
+                var projectDirectoryPath = TrimProject(projectFilePath);
+                if (!string.IsNullOrEmpty(projectDirectoryPath))
+                {
+                    var solutionFoldersPath = TrimProjectDirectory(projectDirectoryPath);
+                    if (!string.IsNullOrEmpty(solutionFoldersPath))
+                    {
+                        solutionFolders.AddRange(solutionFoldersPath.Split(Path.DirectorySeparatorChar));
+                    }
+                }
+            }
+
+            return solutionFolders;
+        }
+
+        private static bool IsPathInTreeRootedAtSolutionDirectory(string path)
+        {
+            return !path.StartsWith("..");
+        }
+
+        private static string TrimProject(string path)
+        {
+            return Path.GetDirectoryName(path);
+        }
+
+        private static string TrimProjectDirectory(string path)
+        {
+            return Path.GetDirectoryName(path);
         }
     }
 }
